fix: remove unnecessary '=' in updateBadge condition  - Removed '=' in the condition that checks the length of recipesCart to correctly handle the visibility of the delete button.
This commit fixes a logical issue in the `upDateBadge` function where the condition `if (recipesCartLength.length >= 0)` incorrectly handled the visibility of the `deleteAllBtn`. Since the length will always be `>= 0`, the condition was adjusted by removing the `=` to ensure proper toggle between displaying and hiding the delete button based on the presence of items in the cart. The updated function now correctly removes the `d-none` class when there are items and hides the button when there are none.